Split out main generator file. Configure robot.
